requirements.txt
DateTime==4.3
Django==3.2
django-crispy-forms==1.12.0
gunicorn==20.1.0
Pillow==8.2.0
